updates in lesson page
